db-api/requirements.txt
fastapi==0.70.0
peewee==3.14.8
python-dotenv==0.19.1
uvicorn==0.15.0
psycopg2-binary==2.9.1
